requirements.txt
pypdfium2==4.30.1
pytesseract==0.3.13
Pillow==11.2.1
spacy==3.8.0
SQLAlchemy==2.0.29
mysql-connector-python==9.3.0
python-dotenv==1.1.0
pytest==8.3.5
tabula-py==2.10.0
pandas==2.2.3
PyMuPDF
jpype1
